Give responsive design to side bar
Now, the side bar retracts when the window shrinks, and comes out when
the window grows. Additionally, the side bar can retract and come out
with the use of the side bar button. When the user wants the side bar to
come out while the window is small, the side bar passes over the main
content rather than pushing it to the side. The only issue is that when
the page is loaded while the window is small, there is a quick animation
of the side bar retracting.