Fix issue with side bar when loading small screen
As a reminder, the issue brought up in the previous commit was that when
loading the web-page while the window was small, the user would be able
to visibly see (albeit quickly) the retraction of the side bar. I
converted the JavaScript responsible for removing the side bar for small
window sizes and converted it to a media query. This solved the issue.