Convert hard coded dashboard into a js module
The hard coded dashboard has now been made into a JavaScript function
module, which can be found in the dom-manipulation.js file, which is
then imported to the index.js file. The css styling has also been made
into a non temporary css file located in the src directory. The event
listeners that deal with the side bar of the dashboard have been placed
in the same module function, however, it may be added to the logic.js
file in the future.