Separate dashboard responsibilities
Initially, the logic of finding out whether the window is shrinking or
growing was included in the dashboard function. This was now made into
its own module pattern function, which can be found in the logic.js
file. Additionally, the dom functions for manipulating the side bar due
to window resize and button click, have now been taken out of the
dashboard function, and made into their own functions (also found in the
dom-manipulation.js file). Finally, event listeners for window resize
and for button clicking have been moved to the index.js file, which
serve as a control of flow for the logic and dom manipulation of the
dashboard side bar.