Design modal window for task form
Designed this modal form window in my inline html and css. The form
is wrapped in a div with class overlay that prevents interaction with
the rest of the UI. Class "is-blurred" was given to each major UI
element, so that an additional blurred filter can be given to the rest
of the UI when the modal window is activated. The next step now will be
to convert this to js and organize code into appropriate files, but not
before first designing and building functionality for adding and
deleting new task categories.