Design add task category functionality
For the time being, I placed the in-line form modal window code in a
temporary file called dist/index-temp.html. For this commit, I designed
the functionality to add new task categories with the click of the plus
icon. This functionality has, as previously, been designed in-line, and
will be placed into the proper src files when the time comes. A factory
function for the task category objects, and an array to store these
objects have been added as well. The next step will be to design the
ability to edit the category name, color, and the ability to remove the
category.