Design remove task category functionality
Once again, I designed a feature in-line to remove task categories. In
the in-line html, I added a trash icon to each task-category created,
which when clicked, will remove the task-category from the library array
and from the UI. I also added some styling that allows for the expanding
of the icons and title found in the task categories, in order to make it
more obvious to the user that these elements can be interacted with.
Finally, I adjusted the z-indexes of various elements so that the info
bubble remains on top of everything, and I added margin and padding to
the task-category styling. I need to keep all this in mind for when
transferring all of this in-line code to the proper files.