Design and implement task category handling
Task categories can now be added, removed, edited, and the beginnings of
being able to select the task category and the contents of which are
displayed in the main content section (only an add task button appears).
I am now designing and implementing new features right away given that
all code and pseudo-code are organized in all of the files managed by
npm and webpack.