Create task form functions
Each add task button is now given another class with a number that
represents the index of the appropriate task category object. These
buttons have also been given styling in the style.css file. The main
task completed for this commit was the creation of the buildTaskForm
and removeTaskForm functions, which are called from two new event
listeners in the index.js file. Now, the add task form can be called and
removed, and by next commit, I should be able to send the user info to
a logic.js function, which will process the data and add a new task to
the task category.