Setup functionality to add and manipulate tasks
Add task form data is now handled in the index.js file. Various logic
and dom manipulation functions are employed to create the new task and
display it in the UI. Here is a list of actions that can now be taken
in addition to creating new tasks:

1. Displayed dates are formatted by the JavaScript library date-fns,
depending on how far out the due date is from today. Due dates within
less than a week will display as the day of the week for example.

2. Tasks can now be edited when clicked. The same form as the add new
task form appears, and the data is handled differently in this context.

3. Each task card contains a complete task icon which can be clicked to
complete it and play a check mark animation.

4. Side bar previously scrolled with the rest of the main content. Now,
I gave the second grid row a height limit, and an overflow-y of auto for
the main content. Now the side bar remains fixed in the y direction.

5. Now, the task category drop down menu can be used, and tasks can be
moved from one task category to another.

6. When editing a task, all of the task info are displayed automatically
in the form.