Create functions for sorting today and late tasks
Created two arrays in the logic.js file, todayLibrary array and
lateLibrary array. Additionally, I created two functions to sort tasks
found in the taskCategoryLibrary array, into the todayLibrary and
lateLibrary arrays. I also set up event listeners for when the today and
late buttons are clicked, which call upon the sorting functions. The
next step will be to create dom-manipulation.js functions to display
these tasks in the UI when called upon.