Complete functionality of today & late selections
The today and late selections are now complete and display the tasks
that are meant to be displayed in them. Here is a list of the following
additions/fixes for these selections to work properly:

1. Tasks that are displayed in the today or late selections can be
completed, which means that an animation of a check mark appearing
occurs, and the task disappears from that selection and from its
corresponding task category. I added getter and setter methods to the
tasks that allow for retrieving and setting of the task index, which
keeps track of where the task belongs. New functions are made use of to
update these task properties when changed.

2. The today and late selections get highlighted when clicked, and the
highlight from the previously clicked task category gets removed.

3. Functions were created to remove everything from the main content
in order to make room for the new container and content of the today and
late selections.