Remove useless functions from the logic.js file
I decided to rethink the reason for function declarations in the
logic.js file. The index.js file is meant to control the flow of the
application. Even though there is a pattern in the index.js file and
therefor "logic", part of controlling the flow of the application is to
hold event listeners and call the necessary logic functions and object
methods. If a logic.js function is only a line long, and that line
consists of calling an object method only, then it is not necessary. for
this reason, I got rid of the following list of functions found in the
logic.js file, and replaced the calling of them in the index.js file
with the necessary code:

1. editTaskCategoryName
2. editTaskCategoryIconColor
3. removeTaskCategory