Add local storage functionality and clean up code
Now when the user refreshes the application, the tasks and task
categories that were created or completed/deleted remain so. This is
because I added a function to iterate through all task categories and
tasks, take their essential primitive data types, make objects out of
them, and strigify and store them in local storage. I also created
functions to get these locally stored objects, re-create the task and
task category objects by passing the primitive elements from the locally
stored objects into the appropriate factory functions, and place them in
the appropriate arrays.

The original direction I went with was to add individual tasks and task
categories to local storage when they were created, and remove them from
local storage individually when they were completed/deleted. The problem
with this was that each task in local storage is linked to its
respective task category based on the position inside the task category
library array that the task category takes. If a task category filled
with tasks gets deleted, the task categories in local storage need to
change their position depending on if there were any task categories
positioned after the deleted one. But the most challenging aspect of
this are the tasks themselves. There are unfortunately too many cases
that require many functions to properly manage the recorded positioning
of these tasks depending on changes that the user makes. A better
solution would be to link tasks to task categories through the name of
the task category, but this would require me to not allow the same task
category names multiple times. This could be a future potential
improvement.