Add change theme functionality
Now when the change theme button on the top bar is clicked, the theme of
the app changes to dark mode or light mode, depending on what theme it
had before. The function that controls this is found in
dom-manipulation.js, and it is called by an event listener in index.js.